+++ src/models/character/Character.ts
@@ -2,6 +2,7 @@ import { JsonObject, JsonReader } from "config_file.js";
 import StarRail from "../../client/StarRail";
 import CharacterData from "./CharacterData";
 import LightCone from "../light_cone/LightCone";
+import Relic from "../relic/Relic";
 
 /**
  * @en Character
@@ -15,11 +16,15 @@ class Character {
     /**  */
     readonly lightCone: LightCone | null;
     /**  */
+    readonly relics: Relic[];
+    /**  */
     readonly level: number;
     /**  */
     readonly exp: number;
     /**  */
     readonly ascension: number;
+    /**  */
+    readonly eidolons: number;
 
     readonly _data: JsonObject;
 
@@ -36,10 +41,12 @@ class Character {
         this.characterData = new CharacterData(json.getAsNumber("AvatarID"), this.client);
 
         this.lightCone = json.has("EquipmentID", "ID") ? new LightCone(json.getAsJsonObject("EquipmentID"), this.client) : null;
+        this.relics = json.getAsJsonArrayWithDefault([], "RelicList").map(relic => new Relic(relic as JsonObject, this.client));
 
         this.level = json.getAsNumber("Level");
-        this.exp = json.getAsNumber("EXP");
-        this.ascension = json.getAsNumber("Promotion");
+        this.exp = json.getAsNumberWithDefault(0, "EXP");
+        this.ascension = json.getAsNumberWithDefault(0, "Promotion");
+        this.eidolons = json.getAsNumberWithDefault(0, "Rank");
     }
 }