+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.3.0
 - Added StarRail#fetchUser.
+- Renamed RelicSubStat to RelicSubStatData.
+- Renamed RelicMainStat to RelicMainStat.
 # 0.2.1
 - Use adm-zip library instead of unzipper.
 # 0.2.0